+++ Frontend/src/pages/1auth/initial_info.css
