+++ Frontend/src/pages/2sharing/detail_sharing.css
