+++ Frontend/src/pages/2sharing/detail_sharing.js
