+++ Frontend/src/pages/2sharing/index.css
