+++ Frontend/src/pages/2sharing/index.js
