+++ Frontend/src/pages/2sharing/post_sharing.css
