+++ Frontend/src/pages/5mypage/edit_profile.js
@@ -1,10 +1,6 @@
-import './profile_edit.css';
+import './edit_profile.css';
 import {useState} from 'react';
 import {BrowserRouter as Router, Route,  Link} from "react-router-dom";
-
-
-
-
 
 function EditProfile() {
     const [adressFilter,setAdressFilter]= useState("")
