--- Frontend/src/pages/profile/index.css
@@ -1,26 +1,0 @@
-.profile_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 75vh;
-
-}
-
-.profile_container .user_name {
-    text-align: center;
-    font-size: 2.2em;
-    font-weight: bold;
-}
-
-.profile_container .user_profile {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.profile_container .user_profile img {
-    margin-top: 2rem;
-    width: 15rem;
-    height: 15rem;
-    border-radius: 50%;
-}
