--- Frontend/src/pages/test/test.css
@@ -1,3 +1,0 @@
-.test {
-  margin-top: 300px;
-}
